Don't Allow Transformer Pod to Die While there are still files to upload to object stpore (#748)
* Add SIGTERM handler to transformer
This is the signal that the Kubernetes control plan will send to the
pod when we are shutting the transform jobs. We now listen for this
signal and then submit a poison pill message to the object store
uploader to shut it down only after the last file has been uploaded
to the object store.

* Explicitly set the termination grace period for transformer pods.
This is the amount of time Kubernetes control plane will allow the
pod to terminate before forcably killing it. We need this time to
make sure all of the files are uploaded to object store.