Correctly record the time when a dataset is created and update the last_used time (#945)
* Correctly record the time when a dataset is created. Also update the last_used time whenever
the dataset is used in a transform.

The create time had incorrectly been set to 0 and the last_used time set to the current date/time.
We now correclty set the last_updated time to the current time along with last_used. Then any time the
dataset is referenced, we update last_used.

* Make initial dataset last used and last updated consistent for new records